refactor(repl): migrate GUI and timing commands into CommandSets
Move start_FABulator, start_openroad_gui and start_klayout_gui into
GuiCommandSet, and timing_model into TimingCommandSet.

stack-info: PR: https://github.com/FPGA-Research/FABulous/pull/891, branch: KelvinChung2000/stack/5
